% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_intake.R
\name{food_intake}
\alias{food_intake}
\title{Food intake estimation}
\usage{
food_intake(avgmass,minmass,maxmass,efficiency,avgenergy,minenergy,maxenergy,iterations)
}
\arguments{
\item{avgmass}{Average body mass of the predator (grams).}

\item{sdmass}{Standard deviation of the body mass of the predator (grams).}

\item{avgenergy}{Average energy content of the prey (kJ/g).}

\item{sdenergy}{Standard deviation of the energy content of the prey (kJ/g).}

\item{efficiency}{Digestion efficiency of the predator (percentage).}

\item{constanta}{Digestion efficiency of the predator (percentage).}

\item{constantb}{Digestion efficiency of the predator (percentage).}

\item{iterations}{Number of iterations (default 100).}
}
\value{
A value (if one iteration) or vector (if multiple iterations) of prey consumption estimations.
}
\description{
Computes pest consumption estimations based on body mass of predators, digestion efficincy and average energy content of predators.
}
\examples{
food_intake()
food_intake()
}
\references{
Alberdi, A., Gilbert, M.T.P. (2019). A guide to the application of Hill numbers to DNA-based diversity analyses. Molecular Ecology Resources, 19, 804-817.\cr\cr
XXXXXX
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{budget}
\keyword{efficiency}
\keyword{energy}
\keyword{mass}
