% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pest_proportion.R
\name{pest_proportion}
\alias{pest_proportion}
\title{Pest proportion estimation}
\usage{
pest_proportion(counttable,sampleinfo,siteinfo,pestinfo,distribution,iterations)
}
\arguments{
\item{counttable}{Average body mass of the predator (grams).}

\item{sampleinfo}{Standard deviation of the body mass of the predator (grams).}

\item{siteinfo}{Maximum body mass (top 5% percentile) of the predator (grams).}

\item{pestinfo}{Digestion efficiency of the predator (percentage).}

\item{distribution}{Average energy content of the prey (kJ/g).}

\item{iterations}{Number of iterations (default 100).}
}
\value{
A value (if one iteration) or vector (if multiple iterations) of prey consumption estimations.
}
\description{
Computes pest consumption estimations.
}
\examples{
pest_proportion()
pest_proportion()
}
\references{
Alberdi, A., Gilbert, M.T.P. (2019). A guide to the application of Hill numbers to DNA-based diversity analyses. Molecular Ecology Resources, 19, 804-817.\cr\cr
XXXXXX
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{budget}
\keyword{efficiency}
\keyword{energy}
\keyword{mass}
