% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predator_density.R
\name{predator_density}
\alias{predator_density}
\title{Predator population density estimation}
\usage{
predator_density(density,enm,distribution,iterations)
}
\arguments{
\item{density}{Matrix or data frame with average, minimum and maximum population density estimations, and the corresponding geographical coordinates.}

\item{enm}{Distribution map shapefile of the predator.}

\item{distribution}{Geographical extension of the analysis.}

\item{iterations}{Geographical resolution if the analysis.}
}
\value{
A raster (if one iteration) or rasterstack (if multiple iterations) object.
}
\description{
XXXXX
}
\examples{
predator_density()
predator_density()
}
\references{
Alberdi, A., Gilbert, M.T.P. (2019). A guide to the application of Hill numbers to DNA-based diversity analyses. Molecular Ecology Resources, 19, 804-817.\cr\cr
XXXXXX
}
\author{
Antton Alberdi, \email{anttonalberdi@gmail.com}
}
\keyword{XXXXX}
